head, tail, uniq: use long counts to avoid 16-bit overflow
On ia16 int is 16 bits, so counts above 32767 overflowed:
head -n and tail -n/-c/+N parsed with atoi into int, and uniq -c
wrapped its duplicate counter. wc/grep/du/sum/cksum already use long.

Verified on ELKS: yes | head -40000 | wc -l now prints 40000
(was empty), uniq -c prints 40000 (was negative), and
tail +79000c skips correctly (byte-exact output).

+++ elkscmd/minix1/uniq.c
@@ -65,10 +65,10 @@ static int equal(char *s1, char *s2)
   return !strcmp(skip(s1), skip(s2));
 }
 
-static void show(char *line, int count)
+static void show(char *line, long count)
 {
   if (cflag)
-	printf("%4d %s", count, line);
+	printf("%4ld %s", count, line);
   else {
 	if ((uflag && count == 1) || (dflag && count != 1))
 		printf("%s", line);
@@ -78,7 +78,7 @@ static void show(char *line, int count)
 static int uniq(void)
 {
   char *p;
-  int seen;
+  long seen;		/* long: duplicate runs can exceed 16-bit int */
 
   /* Setup */
   prevline = buf1;
